TOOL-1907 - Update build process (#35)
Set up a build pipeline for building images and pushing to ECR

Provide expected image scanning (even though this currently does
not run in k8s so is not technically required)

Update the Dockerfile to use blessed sprout images

To get uwsgi in the python 3.11-slim image we need to install it
with pip which requires a compiler to build it.

I moved the entire venv build step up into the builder image so
that we don't have to leave a full compiler installed in the running
image. Then copy the built /venv to the final image.